Increase range of limit_in_bytes

+++ warden-protocol/lib/warden/protocol/limit_memory.rb
@@ -4,11 +4,11 @@ module Warden
   module Protocol
     class LimitMemoryRequest < BaseRequest
       required :handle, :string, 1
-      optional :limit_in_bytes, :uint32, 2
+      optional :limit_in_bytes, :uint64, 2
     end
 
     class LimitMemoryResponse < BaseResponse
-      optional :limit_in_bytes, :uint32, 1
+      optional :limit_in_bytes, :uint64, 1
     end
   end
 end
